Remove spin_directions. Motivation:
It is better to separate the ground state properties from the parameterisation of the Hamiltonian. The former is volatile, while the latter is assumed constant in most of the use-cases of magnopy. In that regard, the modulus of spin should be stored within the SpinHamiltonian class, as it is the parameter of the system. The directions of spin vectors should not be a part of the SpinHamiltonian and have to be provided as a separate variable when needed.